refs #524
